change error-exception processing
- move error classes definition from hm_algo.py into new module
error.py;
- move error-exception processing to infer.py
- raise ParseError in rename.py when local variable is undefined
- add error class name to error msg and change integration tests
accordingly
