+++ core/type_inference/error.py
@@ -1,0 +1,25 @@
+# ----------------------------------------------------------------
+# Exception types
+
+class InferenceError(Exception):
+    """Raised if the type inference algorithm cannot infer types successfully"""
+
+    def __init__(self, message):
+        self.__message = message
+
+    message = property(lambda self: self.__message)
+
+    def __str__(self):
+        return f"{self.__class__.__name__}:  {self.message}"
+
+
+class ParseError(Exception):
+    """Raised if the type environment supplied for is incomplete"""
+
+    def __init__(self, message):
+        self.__message = message
+
+    message = property(lambda self: self.__message)
+
+    def __str__(self):
+        return f"{self.__class__.__name__}:  {self.message}"
